added platform, fixed engine steam-shared problems, fixed client & server (hl2)

+++ src/public/tier1/checksum_crc.h
@@ -1,4 +1,4 @@
-//========= Copyright Valve Corporation, All rights reserved. ============//
+//========= Copyright © 1996-2005, Valve Corporation, All rights reserved. ============//
 //
 // Purpose: Generic CRC functions
 //
@@ -10,20 +10,20 @@
 #pragma once
 #endif
 
-typedef unsigned int CRC32_t;
+typedef unsigned long CRC32_t;
 
-void CRC32_Init( CRC32_t *pulCRC );
-void CRC32_ProcessBuffer( CRC32_t *pulCRC, const void *p, int len );
-void CRC32_Final( CRC32_t *pulCRC );
-CRC32_t	CRC32_GetTableEntry( unsigned int slot );
+void CRC32_Init(CRC32_t* pulCRC);
+void CRC32_ProcessBuffer(CRC32_t* pulCRC, const void* p, int len);
+void CRC32_Final(CRC32_t* pulCRC);
+CRC32_t	CRC32_GetTableEntry(unsigned int slot);
 
-inline CRC32_t CRC32_ProcessSingleBuffer( const void *p, int len )
+inline CRC32_t CRC32_ProcessSingleBuffer(const void* p, int len)
 {
 	CRC32_t crc;
 
-	CRC32_Init( &crc );
-	CRC32_ProcessBuffer( &crc, p, len );
-	CRC32_Final( &crc );
+	CRC32_Init(&crc);
+	CRC32_ProcessBuffer(&crc, p, len);
+	CRC32_Final(&crc);
 
 	return crc;
 }
